Fixed more warnings
git-svn-id: http://www.openflipper.org/svnrepo/OpenFlipper/branches/Free-Staging@16319 383ad7c9-94d9-4d36-a494-682f7c89f535

+++ Plugin-PoissonReconstruction/PoissonReconstruction/SparseMatrix.h
@@ -115,8 +115,8 @@ public:
 	}
 	T2* operator[]( int t ) { return out[t]; }
 	const T2* operator[]( int t ) const { return out[t]; }
-	int threads( void ) const { return int( out.size() ); }
-	void resize( int threads , int dim )
+	int threads( void ) const { return int(out.size()) ; }
+	void resize( size_t threads , int dim )
 	{
 		if( threads!=out.size() || _dim<dim )
 		{
